requirements.txt
absl-py==0.9.0
apyori==1.1.2
asgiref==3.2.3
astor==0.8.1
bokeh==1.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
Django==3.0.2
django-crispy-forms==1.9.0
django-pandas==0.6.1
djangorestframework==3.11.0
gast==0.2.2
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
idna==2.9
intel-openmp==2019.0
Jinja2==2.11.1
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mkl==2019.0
nltk==3.4.5
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.2.0
packaging==20.3
pandas==0.25.3
Pillow==7.1.2
protobuf==3.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
torch==1.4.0+cpu
torchvision==0.5.0+cpu
tornado==6.0.4
urllib3==1.25.8
Werkzeug==1.0.0
wordcloud==1.6.0
wrapt==1.12.1
gunicorn
django-heroku